requirements.txt
yt-dlp
openai-whisper
spacy
gensim
Flask
